Moving loader types under loader module
