+++ src/modules/loader/actions.ts
@@ -1,5 +1,5 @@
 import createAction from '~/utils/createAction';
-import {LoaderActions} from '~/types/loader';
+import {LoaderActions} from './types';
 
 export const dismissLoader = createAction(LoaderActions.dismiss);
 export const setLoader = createAction(LoaderActions.set);
