+++ src/modules/loader/selectors.ts
@@ -1,5 +1,5 @@
 import {RootReducerI} from '~/types/reducer';
-import {LoaderStateI} from '~/types/loader';
+import {LoaderStateI} from './types';
 
 export const getLoaderState = (state: RootReducerI): LoaderStateI =>
   state.loader;
